Fix contact section
Rearrange the contact photos to make the profiles
have a more mixed order. Also, make the contact
card sizes bigger and more uniform to make the
photos easier to see. In addition, make the
overall page bigger to accomodate the extra space
that the contact cards are taking up. The final
major change was making the contact cards
responsive to avoid overflow of the contents on
the page.